net: renesas: rswitch: implement offload for IPv4 multicast routing
Implemented FIB multicast routing notifiers needed for offloading
multicast traffic. The notifier is triggered when kernel multicast
routing table is changed. Since exact addresses are passed as parameters
and it is possible to define interfaces for forwarding, unlike generic
L3 offloading, the notifier is able to add a rule to the L3 table at once.

Signed-off-by: Leonid Komarianskyi <leonid_komarianskyi@epam.com>